[Lexing] Observe nesting in parameter values
Urgh, what a mess. Luckily, we don't care about syntax validity yet, so we
don't need to bother with correctly recognizing the context-dependent nesting
and escaping rules of all 5 punctuation marks. We just need the correct end
position of each individual instruction/macro parameter, and this bit flag +
priority solution works well for some of the more complex cases in JWasm's
test suite.

(Except for that case where you're supposed to be able to use a backslash to
start a comment, and I don't even understand how TASM/JWasm parse that...)