[Lexing] Start a list of directives that can't be preceded by a symbol name
Fixes the recurring OPTION PROC: in JWasm's test suite, which was previously
misidentified as the start of a procedure named OPTION.
However, that's all I'm going to cover for now as far as syntax compliance is
concerned. The remaining edge cases would require a functioning parser that
understands all the size and pointer types, as well as a multi-pass lexing
approach is concerned.