[Lexer] Recursively expand included files
This implementation actually does what I expected from the INCLUDE directive
to begin with, namely, to add the directory of each included file to the
search path of further INCLUDE directives within that included file.

In most cases, this should eliminate the need to manually specify additional
include directories for large code bases (such as ReC98) - which, of course,
can be done here as well.

Since the paths are traversed in ascending order, starting from the directory
of the main assembly file, this does not break (correct) existing code that
may have multiple files with the same name in different directories.