[Maintenance] Get rid of the return value of parse functions
This seemed quite useful when we only parsed IF statements and little else.
But now, every directive has a type, and conditionals have always been the
only type of directives that are evaluated immediately and aren't kept in the
final list of instructions anyway. We could use that return value for
something much more useful, such as…