[Symbols] Warn if case mapping leads to an ambiguous symbol reference
Turns out that JWasm has a completely different (and in my opinion, worse)
idea of how to implement "case mapping": Instead of converting all symbol
names to uppercase and using that to resolve the symbol, it picks the symbol
that was defined *first* and matches the given name in a case-insensitive
string comparison. Since both MASM and TASM do the former, we continue to do
the same, and print a warning whenever this happens.

Yeah, you can only ever encounter this situation if you first set OPTION
CASEMAP to NONE, define some symbols, then set it back to ALL (or NOTPUBLIC).
It's a weird corner case that hopefully no one will ever encounter, but it's
still possible, and that *was* a parsing question I had…