[Symbols] Correctly interpret constness
Redefinitions with different value types are never allowed, regardless of the
symbol's constness, which means that it only applies to the value itself.
Also, redefinitions of the same symbol with an identical value are generally
possible, with each type having its own rules as to when two values are equal.