[Parser] Write structures to the symbol table at ENDS, not at STRUC/UNION
Benign redefinition of structures not only is a MASM syntax feature that we
should support, but also a neater way to not have these structure redefinition
errors in the second pass over the code.

Ironically though, benign structure redefinition is horribly broken in MASM
itself, only JWasm properly supports it in the first place. TASM, on the other
hand, just doesn't care and doesn't enforce constness on structures in the
first place, which is probably the sanest choice.