[Lexer] Disable nesting of < and > when parsing HLL directives
This is actually less hacky than it might seem - the JWasm test suite
mentions a special tokenize mode for HLL directives as well (see
EXPANS39.ASM). It's still rather ugly, but you can't help it in a syntax
where < and > have three different meanings depending on the context.

The only viable alternative would have been to restrict detection of < and
> to the first and last character within a given nesting level. However,
default arguments for multiline macro parameters are a thing and wouldn't
work anymore with such a system (see EXPANS8.ASM).